+++ src/test/java/tech/jhipster/lite/generator/client/paginationdomain/domain/PaginationDomainModuleFactoryTest.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.generator.client.pagination_domain.domain;
+package tech.jhipster.lite.generator.client.paginationdomain.domain;
 
 import static tech.jhipster.lite.module.domain.JHipsterModulesFixture.*;
 import static tech.jhipster.lite.module.infrastructure.secondary.JHipsterModulesAssertions.*;
